\name{B0}
\alias{B0}
\title{Transient Signal}
\description{Transient signal.
}
\usage{data(B0)
}
\format{A vector containing 1024 observations.
}
\source{
  See discussions in the text of ``Practical Time-Frequency Analysis''.
}
\references{
  Carmona, R. A., W. L. Hwang and B Torresani (1998)
  \emph{Practical Time-Frequency Analysis: Gabor and Wavelet Transforms
    with an Implementation in S}, Academic Press, San Diego.
}
\keyword{datasets}
