\name{back1.000}
\alias{back1.000}
\title{Acoustic Returns}
\description{Acoustic returns from natural underwater clutter.
}
\usage{data(back1.000)
}
\format{A vector containing 7936 observations.
}
\source{
  See discussions in the text of ``Practical Time-Frequency Analysis''.
}
\references{
  Carmona, R. A., W. L. Hwang and B Torresani (1998)
  \emph{Practical Time-Frequency Analysis: Gabor and Wavelet Transforms
    with an Implementation in S}, Academic Press, San Diego.
}
\keyword{datasets}
