\name{check.maxresoln}
\alias{check.maxresoln}
\title{
Verify Maximum Resolution
}
\description{
Stop when \eqn{2^{maxresoln}} is larger than the signal size.
}
\usage{
check.maxresoln(maxresoln, np)
}
\arguments{
\item{maxresoln}{
number of decomposition scales.
}
\item{np}{
signal size.
}}
%\value{}
%\details{}
\references{
See discussions in the text of ``Practical Time-Frequency Analysis''.
}
\seealso{
\code{\link{mw}}, \code{\link{mrecons}}.
}
\keyword{ts}
