\name{click}
\alias{click}
\title{Dolphin Click Data}
\description{Dolphin click data.
}
\usage{data(click)
}
\format{A vector containing 2499 observations.
}
\source{
  See discussions in the text of ``Practical Time-Frequency Analysis''.
}
\references{
  Carmona, R. A., W. L. Hwang and B Torresani (1998)
  \emph{Practical Time-Frequency Analysis: Gabor and Wavelet Transforms
    with an Implementation in S}, Academic Press, San Diego.
}
\keyword{datasets}
