\name{crfview}
\alias{crfview}
\title{
Display chained ridges
}
\description{
displays a family of chained ridges, output of \code{\link{cfamily}}.
}
\usage{
crfview(beemap, twod=TRUE)
}
\arguments{
\item{beemap}{
Family of chained ridges, output of \code{\link{cfamily}}. 
}
\item{twod}{
If set to T, displays the ridges as an image. If set to F,
displays as a series of curves.
}}
%\value{}
%\details{}
\references{
See discussions in the text of \dQuote{Practical Time-Frequency Analysis}.
}
\seealso{
\code{\link{crc}},\code{\link{cfamily}} for crazy climbers and corresponding chaining algorithms.
}
\keyword{ts}
