\name{HOWAREYOU}
\alias{HOWAREYOU}
\title{How Are You?}
\description{Example of speech signal.
}
\usage{data(HOWAREYOU)
}
\format{A vector containing 5151 observations.
}
\source{
  See discussions in the text of ``Practical Time-Frequency Analysis''.
}
\references{
  Carmona, R. A., W. L. Hwang and B Torresani (1998)
  \emph{Practical Time-Frequency Analysis: Gabor and Wavelet Transforms
    with an Implementation in S}, Academic Press, San Diego.
}
\keyword{datasets}
