\name{noisywave}
\alias{noisywave}
\title{Noisy Gravitational Wave}
\description{Noisy gravitational wave.
}
\usage{data(noisywave)
}
\format{A vector containing 8192 observations.
}
\source{
  See discussions in the text of ``Practical Time-Frequency Analysis''.
}
\references{
  Carmona, R. A., W. L. Hwang and B Torresani (1998)
  \emph{Practical Time-Frequency Analysis: Gabor and Wavelet Transforms
    with an Implementation in S}, Academic Press, San Diego.
}
\keyword{datasets}
