\name{SVD}
\alias{SVD}
\title{
Singular Value Decomposition
}
\description{
Computes singular value decomposition of a matrix.
}
\usage{
SVD(a)
}
\arguments{
\item{a}{
input matrix.
}}
\value{
a structure containing the 3 matrices of the singular value
decomposition of the input.
}
\details{
\R interface for Numerical Recipes singular value decomposition routine.
}
\references{
See discussions in the text of \dQuote{Time-Frequency Analysis}.
}
%\seealso{}
\keyword{ts}
