\name{wspec.pl}
\alias{wspec.pl}
\title{Log of Wavelet Spectrum Plot}
\description{
  Displays normalized log of wavelet spectrum.
}
\usage{wspec.pl(wspec, nvoice)
}
\arguments{
  \item{wspec}{
    wavelet spectrum.
  }
  \item{nvoice}{
    number of voices.
  }
}
%\value{}
%\details{}
\references{
  See discussions in the text of \dQuote{Practical Time-Frequency Analysis}.
}
\seealso{\code{\link{hurst.est}}.
}
\keyword{ts}
\keyword{hplot}
