\name{YN}
\alias{YN}
\title{Logarithms of the Prices of Japanese Yen}
\description{Logarithms of the prices of a contract of Japanese yen.
}
\usage{data(YN)
}
\format{A vector containing 500 observations.
}
\source{
  See discussions in the text of ``Practical Time-Frequency Analysis''.
}
\references{
  Carmona, R. A., W. L. Hwang and B Torresani (1998)
  \emph{Practical Time-Frequency Analysis: Gabor and Wavelet Transforms
    with an Implementation in S}, Academic Press, San Diego.
}
\keyword{datasets}
