\name{YNdiff}
\alias{YNdiff}
\title{Daily differences of Japanese Yen}
\description{Daily differences of \code{\link{YN}}.
}
\usage{data(YNdiff)
}
\format{A vector containing 499 observations.
}
\source{
  See discussions in the text of ``Practical Time-Frequency Analysis''.
}
\references{
  Carmona, R. A., W. L. Hwang and B Torresani (1998)
  \emph{Practical Time-Frequency Analysis: Gabor and Wavelet Transforms
    with an Implementation in S}, Academic Press, San Diego.
}
\keyword{datasets}
